* 'master' of gitlab.ayla.com.cn:ccpg-hotel/construction-android:
  1.pact整理
  1.阶段提交

+++ app/src/main/java/com/ayla/hotelsaas/ui/ZigBeeAddActivity.java
@@ -83,7 +83,7 @@ public class ZigBeeAddActivity extends BaseMvpActivity<ZigBeeAddView, ZigBeeAddP
         Log.d(TAG, "startBind: " + Thread.currentThread().getName());
         mPresenter.bindZigBeeNodeWithGatewayDSN(
                 getIntent().getStringExtra("deviceId"),
-                getIntent().getLongExtra("cuId", 0),
+                getIntent().getIntExtra("cuId", 0),
                 getIntent().getLongExtra("scopeId", 0),
                 getIntent().getStringExtra("deviceName"));
     }
